backend/requirements.txt
# ============================================================================
# MINIMAL REQUIREMENTS FOR ASSEMBLYAI-ONLY SETUP
# This avoids heavy SpeechBrain dependencies
# ============================================================================

# FastAPI and Server
fastapi==0.119.0
uvicorn[standard]==0.38.0
python-dotenv==1.1.1
aiofiles==25.1.0
pydantic==2.12.3
starlette==0.48.0
python-multipart==0.0.20

# Core Audio Processing (CPU-Only PyTorch)
torch==2.7.0
torchaudio==2.7.0
numpy==1.26.4
soundfile==0.13.1
mutagen==1.47.0

# Audio Enhancement (Pedalboard DRC - Dynamic Range Compression)
pedalboard==0.9.9

# Transcription API (AssemblyAI does transcription + diarization)
assemblyai==0.36.0

# LLM API (Gemini - no local model needed)
google-generativeai==0.8.5

# Database (SQLAlchemy + PostgreSQL)
sqlalchemy==2.0.31
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
argon2-cffi
httpx==0.27.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# ============================================================================
# OPTIONAL: Only if you want Groq as backup
groq==0.32.0
# ============================================================================